ContentDialog fixed, hardcoded strings added
