--- common/Environments/CustomControls/OptimizeDevDriveDialog.xaml.cs
@@ -1,27 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Controls;
-
-namespace DevHome.Common.Environments.CustomControls;
-
-public partial class OptimizeDevDriveDialog : UserControl
-{
-    public OptimizeDevDriveDialog()
-    {
-        InitializeComponent();
-    }
-
-    private void OKButton_Click(object sender, RoutedEventArgs e)
-    {
-        // Handle OK button click here
-        // Close the popup or perform other actions
-    }
-
-    private void CancelButton_Click(object sender, RoutedEventArgs e)
-    {
-        // Handle Cancel button click here
-        // Close the popup or perform other actions
-    }
-}
