+++ src/Styles/FontFamilies.xaml
@@ -6,4 +6,6 @@
          https://github.com/microsoft/cascadia-code/issues/55-->
     <FontFamily x:Key="CascadiaMonoFontFamily">ms-appx:///Assets/Fonts/CascadiaMono.ttf#Cascadia Mono,MS Gothic, NSimSun</FontFamily>
     <FontFamily x:Key="DevHomeFluentIcons">ms-appx:///Assets/Fonts/DevHome.ttf#DevHome Fluent Icons</FontFamily>
+    <!-- Needed for icons on Windows Customization -> Dev Drvie Insights page -->
+    <FontFamily x:Key="AmcFluentIcons">ms-appx:///Assets/Fonts/AmcIcons.ttf#Amc Fluent Icons</FontFamily>
 </ResourceDictionary>
