+++ tools/Customization/DevHome.Customization/CustomControls/DevDriveOptimizerCardBody.xaml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <UserControl
-    x:Class="DevHome.Common.Environments.CustomControls.DevDriveOptimizerCardBody"
+    x:Class="DevHome.Customization.CustomControls.DevDriveOptimizerCardBody"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:devEnvConverters="using:DevHome.Common.Environments.Converters" 
-    xmlns:local="using:DevHome.Common.Environments.CustomControls"
+    xmlns:local="using:DevHome.Customization.CustomControls"
     mc:Ignorable="d">
     <UserControl.Resources>
         <devEnvConverters:CardStateColorToBrushConverter x:Key="CardStateColorToBrushConverter"/>
@@ -33,17 +33,21 @@
                     Text="{x:Bind CacheToBeMoved, Mode=OneWay}"
                     IsTextSelectionEnabled="True"/>
                 <TextBlock Margin="50 10 10 10"
-                    Style="{StaticResource CardBodyStackPanelTextBlockStyle}"
+                    Style="{StaticResource CardBodyStackPanelAltTextBlockStyle}"
                     Text="{x:Bind OptimizationDescription, Mode=OneWay}"
-                    IsTextSelectionEnabled="True"/>
+                    IsTextSelectionEnabled="True"
+                    TextWrapping="Wrap"/>
                 <Button Margin="50 10 10 10"
                     Padding="25 4 25 6"
                     MinWidth="120"
                     Content="Make the changes"
-                    Click="OpenPopup_Click" />
+                    Command="{x:Bind OptimizeDevDriveCommand}" >
+                </Button>
+                <!--   Click="OpenPopup_Click" />
                 <Popup Name="popup" IsOpen="False" >
                     <local:OptimizeDevDriveDialog/>
                 </Popup>
+                -->
             </StackPanel>
         </StackPanel>
         <!--
